requirements.txt
asgiref==3.4.0
astroid==2.8.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.1.1
drf-yasg==1.20.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
Pillow==8.2.0
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.11.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
